refactor out email_aliases_item
